chore: refactor 'mod task' out of state
